agent/pyproject.toml
[project]
name = "autoscaling"
version = "0.1.0"
description = "Project for research at ICN Laboratory"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "influxdb-client>=1.49.0",
    "ipykernel>=6.29.5",
    "kubernetes>=33.1.0",
    "numpy>=2.1.3",
    "prometheus-api-client>=0.6.0",
    "prometheus-client>=0.22.1",
    "python-dotenv>=1.1.1",
    "torch>=2.7.1",
]
